added new quiz - FallingRocks

+++ TheGame/TheGame/BoardInterfaces/IGame.cs
@@ -2,6 +2,6 @@
 {
     public interface IGame
     {
-        int StartGame();
+        double StartGame();
     }
 }
